Fix: xattr reader: read the header after seaking to an OOL value
If an xattr value is stored OOL, the value actually holds an 8 byte
reference to another, previously stored value. This reference points
to the header that we need to read to know the actual size of the
value before reading it, not the value itself, so after reading the
reference and seeking to it, the xattr reader needs to read the actual
header.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@sigma-star.at>